+++ src/agents/workflows/orchestrator/tools.ts
@@ -1,11 +1,5 @@
 import { createTwitterWorkflowTool } from './workflowTools/twitterWorkflowTool.js';
-import {
-  createFetchMentionsTool,
-  createFetchMyRecentRepliesTool,
-  createFetchMyRecentTweetsTool,
-  createFetchTimelineTool,
-  createPostTweetTool,
-} from '../../tools/twitterTools.js';
+import { createAllTwitterTools } from '../../tools/twitterTools.js';
 import { createVectorDbInsertTool, createVectorDbSearchTool } from '../../tools/vectorDbTools.js';
 import { TwitterApi } from '../../../services/twitter/types.js';
 import { VectorDB } from '../../../services/vectorDb/VectorDB.js';
@@ -13,32 +7,21 @@ import { createSaveExperienceTool } from '../../tools/saveExperienceTool.js';
 
 export const createTools = (twitterApi: TwitterApi, vectorDb: VectorDB) => {
   const twitterWorkflowTool = createTwitterWorkflowTool();
-  const fetchMentionsTool = createFetchMentionsTool(twitterApi);
-  const fetchTimelineTool = createFetchTimelineTool(twitterApi);
-  const postTweetTool = createPostTweetTool(twitterApi);
-  const fetchMyRecentRepliesTool = createFetchMyRecentRepliesTool(twitterApi);
-  const fetchMyRecentTweetsTool = createFetchMyRecentTweetsTool(twitterApi);
+  const twitterTools = createAllTwitterTools(twitterApi);
   const vectorDbSearchTool = createVectorDbSearchTool(vectorDb);
   const vectorDbInsertTool = createVectorDbInsertTool(vectorDb);
   const saveExperienceTool = createSaveExperienceTool();
 
   return {
+    ...twitterTools,
+
     twitterWorkflowTool,
-    fetchMentionsTool,
-    fetchTimelineTool,
-    postTweetTool,
-    fetchMyRecentRepliesTool,
-    fetchMyRecentTweetsTool,
     vectorDbSearchTool,
     vectorDbInsertTool,
     saveExperienceTool,
     tools: [
       twitterWorkflowTool,
-      fetchMentionsTool,
-      fetchTimelineTool,
-      postTweetTool,
-      fetchMyRecentRepliesTool,
-      fetchMyRecentTweetsTool,
+      ...twitterTools.tools,
       vectorDbSearchTool,
       vectorDbInsertTool,
       saveExperienceTool,
